issue-65: refactor, extract renderer to an external renderer class
