+++ src/main/java/net/bootsfaces/component/commandButton/CommandButton.java
@@ -1,0 +1,84 @@
+/**
+ *  Copyright 2014 Riccardo Massera (TheCoder4.Eu)
+ *  
+ *  This file is part of BootsFaces.
+ *  
+ *  BootsFaces is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Lesser General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  BootsFaces is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public License
+ *  along with BootsFaces. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package net.bootsfaces.component.commandButton;
+
+import java.io.IOException;
+import java.util.Map;
+
+import javax.faces.FacesException;
+import javax.faces.application.ResourceDependencies;
+import javax.faces.application.ResourceDependency;
+import javax.faces.component.FacesComponent;
+import javax.faces.component.html.HtmlCommandButton;
+import javax.faces.context.FacesContext;
+import javax.faces.context.ResponseWriter;
+import javax.faces.event.ActionEvent;
+
+import net.bootsfaces.listeners.AddResourcesListener;
+import net.bootsfaces.render.A;
+import net.bootsfaces.C;
+import net.bootsfaces.render.H;
+import net.bootsfaces.render.R;
+import net.bootsfaces.render.Tooltip;
+
+/**
+ *
+ * @author thecoder4.eu
+ */
+
+@ResourceDependencies({ @ResourceDependency(library = "bsf", name = "css/core.css", target = "head"),
+		@ResourceDependency(library = "javax.faces", name = "jsf.js", target = "body"),
+		@ResourceDependency(library = "bsf", name = "js/bsf.js", target = "head") })
+@FacesComponent(CommandButton.COMPONENT_TYPE)
+public class CommandButton extends HtmlCommandButton {
+
+	/**
+	 * <p>
+	 * The standard component type for this component.
+	 * </p>
+	 */
+	public static final String COMPONENT_TYPE = "net.bootsfaces.component.commandButton.CommandButton";
+	/**
+	 * <p>
+	 * The component family for this component.
+	 * </p>
+	 */
+	public static final String COMPONENT_FAMILY = "net.bootsfaces.component";
+	
+	/**
+	 * <p>
+	 * Default renderer.
+	 * </p>
+	 */
+	public static final String DEFAULT_RENDERER = "net.bootsfaces.component.commandButton.CommandButton";
+
+
+	public CommandButton() {
+		setRendererType(DEFAULT_RENDERER); // this component renders itself
+		AddResourcesListener.addResourceToHeadButAfterJQuery(C.BSF_LIBRARY, "jq/jquery.js");
+		Tooltip.addResourceFile();
+	}
+
+	@Override
+	public String getFamily() {
+		return COMPONENT_FAMILY;
+	}
+
+}
